Remove the intermediate reporter
In addition, make error submodule export an object that can be configured, rather than a function that has to be called. Furthermore, by default use the project name as a tag. Allow adding tags for individual reportings, as well as overriding the default tags entirely.

Implement slight magic with promises so that Raygun call can safely throw.